refactor profile watcher
