+++ src/app/hooks/use-profile-synchronizer.ts
@@ -2,7 +2,7 @@
 import { isEqual } from "@ardenthq/sdk-helpers";
 import { Contracts } from "@ardenthq/sdk-profiles";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { matchPath, useHistory, useLocation } from "react-router-dom";
+import { useHistory, useLocation } from "react-router-dom";
 import { Services } from "@ardenthq/sdk";
 import { usePrevious } from "./use-previous";
 import { useSynchronizer } from "./use-synchronizer";
@@ -11,8 +11,9 @@ import { DashboardConfiguration } from "@/domains/dashboard/pages/Dashboard";
 import { useAccentColor } from "@/app/hooks/use-accent-color";
 import { useConfiguration, useEnvironmentContext } from "@/app/contexts";
 import { useAutoSignOut } from "@/app/hooks/use-auto-signout";
+import { useProfileWatcher } from "@/app/hooks/use-profile-watcher";
 import { delay } from "@/utils/delay";
-import { getErroredNetworks, getProfileById, getProfileFromUrl, getProfileStoredPassword } from "@/utils/profile-utils";
+import { getErroredNetworks, getProfileFromUrl, getProfileStoredPassword } from "@/utils/profile-utils";
 import { ProfilePeers } from "@/utils/profile-peers";
 import { enabledNetworksCount, profileAllEnabledNetworks, profileEnabledNetworkIds } from "@/utils/network-utils";
 import { useZendesk } from "@/app/contexts/Zendesk";
@@ -24,19 +25,6 @@ enum Intervals {
 	Long = 120_000,
 	VeryLong = 7_200_000,
 }
-
-export const useProfileWatcher = () => {
-	const location = useLocation();
-
-	const { env } = useEnvironmentContext();
-
-	const pathname = (location as any).location?.pathname || location.pathname;
-	const match = useMemo(() => matchPath(pathname, { path: "/profiles/:profileId" }), [pathname]);
-	const profileId = (match?.params as any)?.profileId;
-	const allProfilesCount = env.profiles().count();
-
-	return useMemo(() => getProfileById(env, profileId), [profileId, env, allProfilesCount]); // eslint-disable-line react-hooks/exhaustive-deps
-};
 
 export const useProfileJobs = (profile?: Contracts.IProfile): Record<string, any> => {
 	const { env } = useEnvironmentContext();
